requirements.txt
fastapi
uvicorn[standard]==0.27.1
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
debugpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
httpx
pytest
pytest-asyncio
faker
coverage
asyncpg>=0.27
bcrypt==4.0.1

